chore: forge fmt

+++ src/interfaces/IWETH.sol
@@ -8,5 +8,5 @@ interface IWETH {
 
     function deposit() external payable;
 
-    function transfer(address dst, uint wad) external returns (bool);
+    function transfer(address dst, uint256 wad) external returns (bool);
 }
